Fix task 13 test errors

+++ 0x03-python-data_structures/13-is_palindrome.c
@@ -20,7 +20,7 @@ int is_palindrome(listint_t **head)
 	for (list_len = 1; list->next != NULL; list_len++)
 		list = list->next;
 
-	midpoint = list_len / 2;
+	midpoint = (int)(list_len / 2);
 	--list_len;
 	list = *head;
